requirements.txt
faiss_cpu==1.8.0
faiss_gpu==1.8.0
jax==0.4.33
nltk==3.9.1
numpy==2.1.1
openai==1.50.2
openai_clip==1.0.1
pandas==2.2.3
Pillow==10.4.0
pytorch_ignite==0.5.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
scipy==1.14.1
sentence_transformers==3.1.1
skscope==0.1.8
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
